prod/requirements.txt
azureml-core
azureml-sdk
requests
joblib
